Add files via upload
The goal of this project is to check the effect of different factors on serum creatinine levels over time. Woman with high phenacetin taking has shown a greater change in serum-creatinine level compared with a woman with low phenacetin intake as an estimated parameter is -1.174e-01. So, high phenacetin taker has increased creatinine level over the year, which indicates poor kidney function.